Reorganized source files in core dir into separate libs
