+++ core/decode/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(decode
+  Decode.cpp
+  FusionDecode.cpp
+  MavisUnit.cpp
+)
