+++ core/decode/FusionDecode.cpp
