+++ core/dispatch/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(dispatch
+  Dispatch.cpp
+  Dispatcher.cpp
+)
