+++ core/execute/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(execute
+  Execute.cpp
+  ExecutePipe.cpp
+  IssueQueue.cpp
+)
