+++ core/fetch/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(fetch
+  Fetch.cpp
+  ICache.cpp
+  SimpleBranchPred.cpp
+)
