+++ core/fetch/ICache.cpp
