+++ core/fetch/SimpleBranchPred.cpp
