+++ core/lsu/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(lsu
+  DCache.cpp
+  LSU.cpp
+)
