+++ core/rename/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(rename
+    Rename.cpp
+)
