+++ core/vector/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(vector
+  VectorUopGenerator.cpp
+)
