+++ core/vector/VectorConfig.hpp
